backend/requirements.txt
# 后端 FastAPI + LightRAG
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
lightrag-hku>=0.1.0
numpy>=1.24.0
openai>=1.0.0
aiohttp>=3.9.0
